Create `UiKitContent`
